feat: allow overriding title page elements
The title page elements can now be change to custom values independent of
the metadata set in the `#show` function.
This also allows to selectively not show an item.
For instance, setting `subtitle: none` will not show the subtitle on the
title page, but if set in the metadata, the value is still used for the
footer.